Move the Multi operation trait into its own module
